refactoring: - excluded currentOffer from offers store
